🏗️ Move user related store files to new dedicated feature folder
In the new feature-based architecture, each feature will have its own `store/` folder.
