+++ src/components/App/App.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import { useDispatch } from 'react-redux'
-import { fetchUser } from '../../redux/thunks/user'
+import { fetchUser } from '../../features/user/store/userThunks'
 import { Router } from '../../routes/Router'
 import { LoadingLayout } from '../../layouts/LoadingLayout'
 import { ErrorPageHandler } from '../../routes/ErrorPageHandler'
