+++ src/components/Buttons/ReactionButton/index.jsx
@@ -5,7 +5,7 @@ import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
 import { useDispatch, useSelector } from 'react-redux'
 import { getPostReactions } from '../../../redux/selectors/feed'
-import { getUserId } from '../../../redux/selectors/user'
+import { getUserId } from '../../../features/user/store/userSelectors'
 import { createReaction, updateReaction, removeReaction } from '../../../redux/thunks/feed'
 import './style.scss'
