+++ src/components/Feed/index.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { Box, Typography, Avatar, Grid, CircularProgress } from '@mui/material'
-import { getUser, getUserOrganizationName } from '../../redux/selectors/user'
+import { getUser, getUserOrganizationName } from '../../features/user/store/userSelectors'
 import { getPosts, getHasMorePosts, getPostLoading } from '../../redux/selectors/feed'
 import { fetchPosts } from '../../redux/thunks/feed'
 import { cleanFeedState } from '../../redux/reducers/feed'
