+++ src/components/Post/index.jsx
@@ -9,7 +9,7 @@ import { HashLink } from 'react-router-hash-link'
 import pluralize from 'pluralize'
 import { getPostComments, getPostReactions } from '../../redux/selectors/feed'
 import { fetchComments } from '../../redux/thunks/feed'
-import { getUser } from '../../redux/selectors/user'
+import { getUser } from '../../features/user/store/userSelectors'
 import { CommentForm } from '../Forms/CommentForm'
 import { Comment } from '../Comment'
 import { ReactionButton } from '../Buttons/ReactionButton'
