+++ src/features/user/components/ProfileForm/index.jsx
@@ -4,8 +4,8 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useForm } from 'react-hook-form'
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom'
 import { useEffect, useState } from 'react'
-import { getUser, getIsLogged } from '../../../../redux/selectors/user'
-import { createUser, updateUser } from '../../../../redux/reducers/user'
+import { getUser, getIsLogged } from '../../store/userSelectors'
+import { createUser, updateUser } from '../../store/userSlice'
 import { AvatarFormInput } from '../AvatarFormInput'
 import { api, fetchCsrfCookie } from '../../../../services/api'
 import { useServerErrors } from '../../../../hooks/useServerErrors'
