+++ src/routes/useAxiosInterceptors.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { useDispatch } from 'react-redux'
 import { api } from '../services/api'
-import { cleanUserState } from '../redux/reducers/user'
+import { cleanUserState } from '../features/user/store/userSlice'
 
 export function useAxiosInterceptors() {
     const dispatch = useDispatch()
